Divided preferences into several fragments, each with own layout. Closes #692
