+++ app/src/main/java/org/kontalk/ui/PreferencesActivity.java
@@ -38,6 +38,12 @@ import android.widget.Toast;
 import org.kontalk.R;
 import org.kontalk.authenticator.Authenticator;
 import org.kontalk.crypto.PersonalKeyPack;
+import org.kontalk.ui.prefs.AppearanceFragment;
+import org.kontalk.ui.prefs.MaintenanceFragment;
+import org.kontalk.ui.prefs.MediaFragment;
+import org.kontalk.ui.prefs.MessagingFragment;
+import org.kontalk.ui.prefs.NetworkFragment;
+import org.kontalk.ui.prefs.NotificationFragment;
 import org.kontalk.util.Preferences;
 
 
@@ -47,7 +53,7 @@ import org.kontalk.util.Preferences;
  * @author Daniele Ricci
  */
 public final class PreferencesActivity extends ToolbarActivity
-        implements PreferencesFragment.Callback, FolderChooserDialog.FolderCallback, ColorChooserDialog.ColorCallback {
+        implements RootPreferenceFragment.Callback, FolderChooserDialog.FolderCallback, ColorChooserDialog.ColorCallback {
     private static final String TAG_NESTED = "nested";
 
     @Override
@@ -99,8 +105,26 @@ public final class PreferencesActivity extends ToolbarActivity
     public void onNestedPreferenceSelected(int key) {
         Fragment fragment;
         switch (key) {
+            case R.xml.preferences_network:
+                fragment = new NetworkFragment();
+                break;
+            case R.xml.preferences_messaging:
+                fragment = new MessagingFragment();
+                break;
             case R.xml.privacy_preferences:
                 fragment = new PrivacyPreferences();
+                break;
+            case R.xml.preferences_appearance:
+                fragment = new AppearanceFragment();
+                break;
+            case R.xml.preferences_media:
+                fragment = new MediaFragment();
+                break;
+            case R.xml.preferences_notification:
+                fragment = new NotificationFragment();
+                break;
+            case R.xml.preferences_maintenance:
+                fragment = new MaintenanceFragment();
                 break;
             default:
                 throw new IllegalArgumentException("unknown preference screen: " + key);
@@ -114,8 +138,8 @@ public final class PreferencesActivity extends ToolbarActivity
     @Override
     public void onFolderSelection(@NonNull FolderChooserDialog folderChooserDialog, @NonNull File folder) {
         try {
-            PreferencesFragment f = (PreferencesFragment) getSupportFragmentManager()
-                .findFragmentById(R.id.container);
+            MaintenanceFragment f = (MaintenanceFragment) getSupportFragmentManager()
+                    .findFragmentById(R.id.container);
 
             f.exportPersonalKey(this,
                 new FileOutputStream(new File(folder, PersonalKeyPack.KEYPACK_FILENAME)));
