+++ app/src/main/java/org/kontalk/ui/RootPreferenceFragment.java
@@ -21,10 +21,13 @@ package org.kontalk.ui;
 import com.github.machinarius.preferencefragment.PreferenceFragment;
 
 import android.app.Activity;
+import android.content.Intent;
 import android.os.Bundle;
 import android.preference.Preference;
+import android.view.MenuItem;
 
 import org.kontalk.R;
+import org.kontalk.ui.ConversationsActivity;
 
 
 /**
@@ -58,22 +61,23 @@ public class RootPreferenceFragment extends PreferenceFragment {
         mCallback = null;
     }
 
-    void invokeCallback(int key) {
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                getActivity().onBackPressed();
+                return true;
+        }
+
+        return super.onOptionsItemSelected(item);
+    }
+
+    public void invokeCallback(int key) {
         if (mCallback != null)
             mCallback.onNestedPreferenceSelected(key);
     }
 
-    protected void setupPreferences() {
-        // privacy section
-        final Preference privacy = findPreference("pref_privacy_settings");
-        privacy.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-            @Override
-            public boolean onPreferenceClick(Preference preference) {
-                invokeCallback(R.xml.privacy_preferences);
-                return true;
-            }
-        });
-    }
+    protected void setupPreferences() {}
 
     public interface Callback {
         public void onNestedPreferenceSelected(int key);
