+++ app/src/main/java/org/kontalk/ui/prefs/MediaFragment.java
@@ -1,0 +1,48 @@
+/*
+ * Kontalk Android client
+ * Copyright (C) 2016 Kontalk Devteam <devteam@kontalk.org>
+
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.kontalk.ui.prefs;
+
+import android.os.Bundle;
+
+import org.kontalk.R;
+import org.kontalk.ui.PreferencesActivity;
+import org.kontalk.ui.RootPreferenceFragment;
+
+/**
+ * Messaging settings fragment.
+ */
+public class MediaFragment extends RootPreferenceFragment {
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        // Load the preferences from an XML resource
+        addPreferencesFromResource(R.xml.preferences_media);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+
+        ((PreferencesActivity) getActivity()).getSupportActionBar()
+                .setTitle(R.string.pref_media_settings);
+    }
+
+}
