+++ app/src/main/res/xml/preferences_appearance.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Kontalk Android client
+     Copyright (C) 2016 Kontalk Devteam <devteam@kontalk.org>
+
+     This program is free software: you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation, either version 3 of the License, or
+     (at your option) any later version.
+
+     This program is distributed in the hope that it will be useful,
+     but WITHOUT ANY WARRANTY; without even the implied warranty of
+     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     GNU General Public License for more details.
+
+     You should have received a copy of the GNU General Public License
+     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ -->
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <com.afollestad.materialdialogs.prefs.MaterialListPreference
+        android:key="pref_font_size"
+        android:title="@string/pref_font_size"
+        android:summary="%s"
+        android:entries="@array/fontsize_labels"
+        android:entryValues="@array/fontsize_values"
+        android:defaultValue="@string/pref_default_font_size"/>
+    <CheckBoxPreference android:key="pref_custom_background"
+        android:title="@string/pref_custom_background"
+        android:summary="@string/pref_title_custom_background"
+        android:layout="@layout/md_preference_custom"/>
+    <Preference android:key="pref_background_uri"
+        android:title="@string/pref_background_uri"
+        android:summary="@string/pref_title_background_uri"
+        android:dependency="pref_custom_background"
+        android:layout="@layout/md_preference_custom"/>
+    <com.afollestad.materialdialogs.prefs.MaterialListPreference
+        android:key="pref_balloons"
+        android:title="@string/pref_balloons"
+        android:summary="%s"
+        android:entries="@array/balloons_labels"
+        android:entryValues="@array/balloons_values"
+        android:defaultValue="@string/pref_default_balloons"/>
+    <com.afollestad.materialdialogs.prefs.MaterialListPreference
+        android:key="pref_text_enter"
+        android:title="@string/pref_text_enter"
+        android:summary="@string/pref_title_text_enter"
+        android:entries="@array/text_enter_labels"
+        android:entryValues="@array/text_enter_values"
+        android:defaultValue="@string/pref_default_text_enter"/>
+
+</PreferenceScreen>
