+++ app/src/main/res/xml/preferences_maintenance.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Kontalk Android client
+     Copyright (C) 2016 Kontalk Devteam <devteam@kontalk.org>
+
+     This program is free software: you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation, either version 3 of the License, or
+     (at your option) any later version.
+
+     This program is distributed in the hope that it will be useful,
+     but WITHOUT ANY WARRANTY; without even the implied warranty of
+     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     GNU General Public License for more details.
+
+     You should have received a copy of the GNU General Public License
+     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ -->
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <com.afollestad.materialdialogs.prefs.MaterialListPreference
+        android:key="pref_idle_time"
+        android:title="@string/pref_idle_time"
+        android:summary="@string/pref_title_idle_time"
+        android:entries="@array/idle_time_labels"
+        android:entryValues="@array/idle_time_values"
+        android:defaultValue="@integer/pref_default_idle_time"/>
+    <com.afollestad.materialdialogs.prefs.MaterialListPreference
+        android:key="pref_wakeup_time"
+        android:title="@string/pref_wakeup_time"
+        android:summary="@string/pref_title_wakeup_time"
+        android:entries="@array/wakeup_time_labels"
+        android:entryValues="@array/wakeup_time_values"
+        android:defaultValue="@integer/pref_default_wakeup_time"/>
+    <Preference android:key="pref_restart_msgcenter"
+        android:summary="@string/pref_title_restart_msgcenter"
+        android:title="@string/pref_restart_msgcenter"
+        android:layout="@layout/md_preference_custom"/>
+    <Preference android:key="pref_change_passphrase"
+        android:title="@string/pref_change_passphrase"
+        android:layout="@layout/md_preference_custom"/>
+    <Preference android:key="pref_regenerate_keypair"
+        android:summary="@string/pref_title_regenerate_keypair"
+        android:title="@string/pref_regenerate_keypair"
+        android:layout="@layout/md_preference_custom"/>
+    <Preference android:key="pref_export_keypair"
+        android:title="@string/pref_export_keypair"
+        android:layout="@layout/md_preference_custom"/>
+    <CheckBoxPreference android:key="pref_accept_any_certificate"
+        android:title="@string/pref_accept_any_certificate"
+        android:defaultValue="@bool/pref_default_accept_any_certificate"
+        android:layout="@layout/md_preference_custom"/>
+    <Preference android:key="pref_delete_account"
+        android:title="@string/pref_delete_account"
+        android:layout="@layout/md_preference_custom"/>
+
+</PreferenceScreen>
