+++ app/src/main/res/xml/preferences_messaging.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Kontalk Android client
+     Copyright (C) 2016 Kontalk Devteam <devteam@kontalk.org>
+
+     This program is free software: you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation, either version 3 of the License, or
+     (at your option) any later version.
+
+     This program is distributed in the hope that it will be useful,
+     but WITHOUT ANY WARRANTY; without even the implied warranty of
+     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     GNU General Public License for more details.
+
+     You should have received a copy of the GNU General Public License
+     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ -->
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <Preference android:key="pref_privacy_settings"
+        android:title="@string/pref_privacy_settings"
+        android:persistent="false"
+        android:layout="@layout/md_preference_custom"/>
+    <CheckBoxPreference android:key="pref_show_blocked_users"
+        android:title="@string/pref_show_blocked_users"
+        android:defaultValue="@bool/pref_default_show_blocked_users"
+        android:layout="@layout/md_preference_custom"/>
+    <com.afollestad.materialdialogs.prefs.MaterialEditTextPreference
+        android:key="pref_remove_prefix"
+        android:dialogTitle="@string/pref_remove_prefix"
+        android:positiveButtonText="@android:string/ok"
+        android:negativeButtonText="@android:string/cancel"
+        android:title="@string/pref_remove_prefix"
+        android:summary="@string/pref_title_remove_prefix"
+        android:singleLine="true" android:inputType="phone"/>
+    <CheckBoxPreference android:key="pref_sync_sim_contacts"
+        android:summary="@string/pref_title_sync_sim_contacts"
+        android:title="@string/pref_sync_sim_contacts"
+        android:defaultValue="@bool/pref_default_sync_sim_contacts"
+        android:layout="@layout/md_preference_custom"/>
+    <CheckBoxPreference android:key="pref_sync_invisible_contacts"
+        android:summary="@string/pref_title_sync_invisible_contacts"
+        android:title="@string/pref_sync_invisible_contacts"
+        android:defaultValue="@bool/pref_default_sync_invisible_contacts"
+        android:layout="@layout/md_preference_custom"/>
+
+</PreferenceScreen>
