+++ app/src/main/res/xml/preferences_network.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Kontalk Android client
+     Copyright (C) 2016 Kontalk Devteam <devteam@kontalk.org>
+
+     This program is free software: you can redistribute it and/or modify
+     it under the terms of the GNU General Public License as published by
+     the Free Software Foundation, either version 3 of the License, or
+     (at your option) any later version.
+
+     This program is distributed in the hope that it will be useful,
+     but WITHOUT ANY WARRANTY; without even the implied warranty of
+     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+     GNU General Public License for more details.
+
+     You should have received a copy of the GNU General Public License
+     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ -->
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <Preference android:key="pref_update_server_list"
+        android:title="@string/pref_update_server_list"
+        android:layout="@layout/md_preference_custom"/>
+    <com.afollestad.materialdialogs.prefs.MaterialEditTextPreference
+        android:key="pref_network_uri"
+        android:dialogTitle="@string/pref_network_uri"
+        android:positiveButtonText="@android:string/ok"
+        android:negativeButtonText="@android:string/cancel"
+        android:title="@string/pref_network_uri"
+        android:summary="@string/pref_title_network_uri"
+        android:singleLine="true" android:inputType="textUri"/>
+    <CheckBoxPreference android:key="pref_push_notifications"
+        android:summary="@string/pref_title_push_notifications"
+        android:title="@string/pref_push_notifications"
+        android:defaultValue="@bool/pref_default_push_notifications"
+        android:layout="@layout/md_preference_custom"/>
+
+</PreferenceScreen>
